+++ R/fix_dates.R
@@ -12,11 +12,15 @@ months$full <- tolower(month.name)
 #' US system of MDY is not supported.
 #' @param df A \code{dataframe} object with messy date column(s)
 #' @param col.names Character vector of names of columns of messy date data
+#' @param id Name of column containing row IDs. By default, the first column is
+#' assumed.
 #' @param day.impute Integer. Day of the month to be imputed if not available.
 #'   defaults to 1.
 #' @param month.impute Integer. Month to be be imputed if not available.
 #'   Defaults to 7 (July)
 #' @return A \code{dataframe} object. Selected columns are of type \code{Date}
+#' @seealso \link{fix_date} Similar to \code{fix_dates()} except can only be
+#' applied to character obkects. 
 #' @examples
 #' bad.dates <- data.frame(id = seq(5),
 #'                         some.dates = c("02/05/92",
@@ -31,31 +35,52 @@ months$full <- tolower(month.name)
 #'                                             "jan 2020"))
 #'fixed.df <- fix_dates(bad.dates, c("some.dates", "some.more.dates"))
 #' @export
-fix_dates <- function(df, col.names, day.impute = 1, month.impute = 7) {
+fix_dates <- function(df,
+                      col.names,
+                      day.impute = 1,
+                      month.impute = 7,
+                      id = NULL) {
   if (!is.data.frame(df)) {
     stop("df should be a dataframe object!")
   }
   if (any(!is.character(col.names))) {
     stop("col.names should be a character vector!")
   }
+  
+  if (is.null(id)) id <- 1 # Use first column as id if not explictly given
 
   .checkday(day.impute)
   .checkmonth(month.impute)
   day.impute <- .convertimpute(day.impute)
   month.impute <- .convertimpute(month.impute)
-
-
+  error.status <- 0
+  
   for (col.name in col.names) {
     fixed.dates <- c()
       for (i in 1:nrow(df)) {
-        fixed.dates[i] <- fix_date(df[i, col.name], day.impute, month.impute)
+        tryCatch(
+          {
+            fixed.dates[i] <- .fix_date(df[i, col.name],
+                                       day.impute,
+                                       month.impute,
+                                       subject = df[i, id])
+            },
+          error = function(cond){
+            message(paste0("Unable to resolve date for subject ",
+                       df[i, id],
+                       " (date: ",
+                       df[i, col.name],
+                       ")\n")
+                 )
+            stop(cond)}
+        )
       }
     df[, col.name] <- as.Date(fixed.dates)
     }
     df
 }
 
-fix_date <- function(date, day.impute, month.impute) {
+.fix_date <- function(date, day.impute, month.impute, subject) {
 
   if (is.null(date) || is.na(date) || as.character(date) == "") {
     return(NA)
@@ -65,9 +90,11 @@ fix_date <- function(date, day.impute, month.impute) {
 
   if (nchar(date) == 4) {
     # Just given year
-    year <- date; month <- month.impute; day <- day.impute
+    year <- date
+    month <- .imputemonth(month.impute)
+    day <- .imputeday(day.impute)
   } else{
-    date_vec <- seperate_date(date)
+    date_vec <- .separate_date(date)
     if (any(nchar(date_vec) > 4)) {
       stop("unable to tidy a date")
     }
@@ -94,7 +121,7 @@ fix_date <- function(date, day.impute, month.impute) {
     }
     if (length(date_vec) < 3) {
       # ASSUME MM/YYYY, YYYY/MM
-      day <- day.impute
+      day <- .imputeday(day.impute) 
       if (nchar(date_vec[1]) == 4) {
       # Assume YYYY/MM
       year <- date_vec[1]; month <- date_vec[2]
@@ -111,11 +138,23 @@ fix_date <- function(date, day.impute, month.impute) {
       }
     }
   }
-  fixed_date <- paste0(year, "-", month, "-", day)
+  .checkoutput(day, month)
+  
+  if (is.na(day) || is.na(month)) {
+    fixed_date <- NA
+    warning(paste0("NA imputed for subject ",
+                   subject,
+                   " (date: ",
+                   date,
+                   ")\n"),
+            call. = FALSE)
+  } else { 
+    fixed_date <- paste0(year, "-", month, "-", day)
+  }
   fixed_date
 }
 
-seperate_date <- function(date) {
+.separate_date <- function(date) {
   if (grepl("/", date, fixed = TRUE)) {
     date_vec <- stringr::str_split_fixed(date,
                                          pattern = "/",
@@ -140,10 +179,11 @@ seperate_date <- function(date) {
     } else {
       replacement <- i
     }
-    date <- gsub(pattern = months$abbrev[i],
+    date <- gsub(pattern = months$full[i],
                  replacement = replacement,
                  x = date)
-    date <- gsub(pattern = months$full[i],
+    
+    date <- gsub(pattern = months$abbrev[i],
                  replacement = replacement,
                  x = date)
   }
@@ -151,32 +191,68 @@ seperate_date <- function(date) {
 }
 
 .checkday <- function(day.impute){
-  if (day.impute < 1 | day.impute >28){
-    stop("day.impute should be an integer between 1 and 28\n")
-  }
-  if (!(day.impute%%1==0)){
-    stop("day.impute should be an integer\n")
+  if (!is.na(day.impute) && !is.null(day.impute)) {
+    if (day.impute < 1 | day.impute > 28) {
+      stop("day.impute should be an integer between 1 and 28\n")
+    }
+    if (!(day.impute %% 1 == 0)) {
+      stop("day.impute should be an integer\n")
+    }
   }
   return()
 }
 
 .checkmonth <- function(month.impute){
-  if (month.impute < 1 | month.impute > 12){
-    stop("month.impute should be an integer between 1 and 12\n")
-  }
-  if (!(month.impute%%1==0)){
-    stop("month.impute should be an integer\n")
+  if (!is.na(month.impute) && !is.null(month.impute)) {
+    if (month.impute < 1 | month.impute > 12) {
+      stop("month.impute should be an integer between 1 and 12\n")
+    }
+    if (!(month.impute %% 1 == 0)) {
+      stop("month.impute should be an integer\n")
+    }
   }
   return()
 }
 
+.checkoutput <- function(day, month){
+  if (!is.na(month)) {
+    if (as.numeric(month) > 12 | as.numeric(month) < 1) {
+      stop("Month not in expected range \n")
+    }
+  }
+  if (!is.na(day)) {
+    if (as.numeric(day) > 31 | as.numeric(day) < 1) {
+      stop("Day of the year not in expected range \n")
+    }
+  }
+  NULL
+}
 
 .convertimpute <- function(impute){
-  if (impute < 10) {
-    replacement <- paste0("0", impute)
+  if (!is.na(impute) && !is.null(impute)) {
+    if (impute < 10) {
+      replacement <- paste0("0", impute)
+    } else {
+      replacement <- as.character(impute)
+    }
   } else {
-    replacement <- as.character(impute)
+    replacement <- impute
   }
   replacement
 }
 
+.imputemonth <- function(month.impute){
+  if (is.null(month.impute)) {
+    stop("Missing month with no imputation value given \n")
+  } else {
+    month.impute
+  }
+}
+
+.imputeday <- function(day.impute){
+  if (is.null(day.impute)) {
+    stop("Missing day with no imputation value given \n")
+  } else {
+    day.impute
+  }
+}